refactor: remove circular dependency from useCodeSelection hook
- Hook now only exports state and handlers (no JSX)

- Editors import and render CodeSelectionMenu directly

- Renamed useCodeSelection.tsx to useCodeSelection.ts (no JSX needed)
